fix: address PR review feedback
- Use getServerAuthToken() in page.tsx instead of inline cookie reading
  with wrong order (fixes the cookie inconsistency the PR aimed to fix)
- Extract shared createMockRedis() to helpers/mock-redis.ts, replacing
  duplicate ~250-line implementations in session-persistence and
  queue-sync-fixes tests
- Document Postgres-only fallback mode in websocket-implementation.md
- Simplify baseQuery() in activity-feed.ts (let → implicit return)
